requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
atk-1.0=2.38.0=hd4edc92_1
blas=1.0=mkl
blosc=1.21.5=hc2324a3_1
boost-cpp=1.85.0=h44aadfe_2
bottleneck=1.4.2=py312ha883a20_0
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
brotli-python=1.0.9=py312h6a678d5_9
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.4=hb9d3cd8_0
ca-certificates=2025.1.31=hbcca054_0
cairo=1.16.0=hb05425b_5
certifi=2025.1.31=pyhd8ed1ab_0
cffi=1.17.1=py312h1fdaa30_1
cfitsio=3.470=hb418390_7
charset-normalizer=3.4.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_1
contourpy=1.3.1=py312hdb19cb5_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.1=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.77=0
cuda-runtime=12.4.1=0
cuda-version=12.6=3
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.27=h9033bb2_6
dbus=1.13.18=hb2f20db_0
eigen=3.4.0=h00ab1b0_0
expat=2.6.4=h5888daf_0
ffmpeg=4.2.3=h167e202_0
filelock=3.16.1=pyhd8ed1ab_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.14.1=h55d465d_3
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.55.3=pyhe1237c8_1
freetype=2.12.1=h267a509_2
freexl=2.0.0=h743c826_0
fribidi=1.0.10=h36c2ea0_0
fsspec=2024.12.0=pyhd8ed1ab_0
gdk-pixbuf=2.42.10=h05c8ddd_0
geos=3.10.6=h6a678d5_0
geotiff=1.7.0=h2a26cda_3
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
giflib=5.2.2=hd590300_0
glib=2.78.4=hfc55251_0
glib-tools=2.78.4=hfc55251_0
gmp=6.3.0=hac33072_2
gnutls=3.6.13=h85f3911_1
grad-cam=1.5.4=pyhd8ed1ab_1
graphite2=1.3.14=h295c915_1
graphviz=2.50.0=h8e749b2_2
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
gtk2=2.24.33=h90689f9_2
gts=0.7.6=h977cf35_4
h2=4.1.0=pyhd8ed1ab_1
harfbuzz=4.3.0=hf52aaf7_2
hdf4=4.2.13=h10796ff_1005
hdf5=1.12.1=nompi_h4df4325_104
hpack=4.0.0=pyhd8ed1ab_1
huggingface_hub=0.27.1=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_1
icu=73.2=h59595ed_0
idna=3.10=pyhd8ed1ab_1
imageio=2.37.0=pyhfb79c49_0
importlib-metadata=8.6.1=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.5=pyhd8ed1ab_0
joblib=1.4.2=pyhd8ed1ab_1
jpeg=9e=h166bdaf_2
json-c=0.16=hc379101_0
kealib=1.5.0=hd940352_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.8=py312h6a678d5_0
krb5=1.20.1=h81ceb04_0
lame=3.100=h166bdaf_1003
lazy-loader=0.4=pyhd8ed1ab_2
lazy_loader=0.4=pyhd8ed1ab_2
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.43=h712a8e2_2
lerc=4.0.0=h27087fc_0
libabseil=20240116.2=cxx17_he02047a_1
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libblas=3.9.0=1_h86c2bf4_netlib
libboost=1.85.0=hba137d9_2
libboost-devel=1.85.0=h00ab1b0_2
libboost-headers=1.85.0=ha770c72_2
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcblas=3.9.0=8_h3b12eaf_netlib
libclang=14.0.6=default_h7634d5b_1
libclang13=14.0.6=default_h9986a30_1
libcublas=12.4.5.8=0
libcufft=11.2.1.3=0
libcufile=1.11.1.6=0
libcups=2.3.3=h36d4200_3
libcurand=10.3.7.77=0
libcurl=8.11.1=hc9e6f67_0
libcusolver=11.6.1.9=0
libcusparse=12.3.1.170=0
libdeflate=1.22=hb9d3cd8_0
libedit=3.1.20240808=pl5321h7949ede_0
libev=4.33=hd590300_2
libexpat=2.6.4=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgd=2.3.3=h695aa2c_1
libgdal=3.6.2=he9acc32_9
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libgfortran=14.2.0=h69a702a_1
libgfortran-ng=14.2.0=h69a702a_1
libgfortran5=14.2.0=hd5240d6_1
libglib=2.78.4=h783c2da_0
libgomp=14.2.0=h77fa898_1
libhwloc=2.11.2=default_h0d58e46_1001
libiconv=1.17=hd590300_2
libjpeg-turbo=2.0.0=h9bf148f_0
libkml=1.3.0=h01aab08_1018
liblapack=3.9.0=8_h3b12eaf_netlib
libllvm14=14.0.6=hcd5def8_4
libltdl=2.4.3a=h5888daf_0
liblzma=5.6.3=hb9d3cd8_1
liblzma-devel=5.6.3=hb9d3cd8_1
libnetcdf=4.8.1=h14805e7_4
libnghttp2=1.58.0=h47da74e_1
libnpp=12.2.5.30=0
libnsl=2.0.1=hd590300_0
libntlm=1.8=hb9d3cd8_0
libnvfatbin=12.6.77=0
libnvjitlink=12.4.127=0
libnvjpeg=12.3.1.117=0
libpng=1.6.43=h2797004_0
libpq=17.2=hdbd6064_0
libprotobuf=4.25.3=h08a7969_0
librsvg=2.54.4=h7abd40a_0
libspatialite=5.1.0=h798b768_2
libsqlite=3.46.0=hde9e2c9_0
libssh2=1.11.1=h251f7ec_0
libstdcxx=14.2.0=hc0a3c3a_1
libstdcxx-ng=14.2.0=h4852527_1
libtiff=4.5.1=hffd6297_1
libtool=2.5.4=h5888daf_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=hd590300_1
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h2c5496b_1
libxml2=2.13.5=hfdd30dd_0
libzip=1.10.1=h2629f0a_3
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=15.0.7=h0cdce71_0
lz4-c=1.9.4=hcb278e6_0
markupsafe=3.0.2=pyhe1237c8_1
matplotlib=3.10.0=py312h06a4308_0
matplotlib-base=3.10.0=py312hbfdbfaf_0
minizip=4.0.6=h9d307f2_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py312h5eee18b_2
mkl_fft=1.3.11=py312h5eee18b_0
mkl_random=1.2.8=py312h526ad5a_0
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
munkres=1.1.4=pyh9f0ad1d_0
mysql=8.4.0=h29a9f33_1
ncurses=6.5=h2d0b736_2
nettle=3.6=he412f7d_0
networkx=3.4.2=pyh267e887_2
nspr=4.36=h5888daf_0
nss=3.100=hca3bf56_0
numexpr=2.10.1=py312h3c60e43_0
numpy=1.26.4=py312hc5e2394_0
numpy-base=1.26.4=py312h0da6c21_0
opencv=4.10.0=py312h907dce5_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openldap=2.6.5=h389dcaa_0
openssl=3.4.1=h7b32b05_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=py312h6a678d5_0
pango=1.50.7=hbd2fdc8_0
pcre2=10.42=hcad00b1_0
pillow=11.0.0=py312hcea889d_1
pip=24.3.1=pyh8b19718_2
pixman=0.44.2=h29eaf8c_0
ply=3.11=pyhd8ed1ab_3
poppler=24.09.0=hcf11d46_1
poppler-data=0.4.12=hd8ed1ab_0
proj=9.3.1=he5811b7_0
pthread-stubs=0.4=hb9d3cd8_1002
pycparser=2.22=pyh29332c3_1
pygraphviz=1.14=py312h5eee18b_0
pyparsing=3.2.1=pyhd8ed1ab_0
pyqt=5.15.10=py312h6a678d5_0
pyqt5-sip=12.13.0=py312h5eee18b_0
pysocks=1.7.1=pyha55dd90_7
python=3.12.8=h5148396_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-tzdata=2024.2=pyhd8ed1ab_1
pytorch=2.5.1=py3.12_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_7
pytorch-mutex=1.0=cuda
pytz=2024.1=pyhd8ed1ab_0
pyyaml=6.0.2=py312h5eee18b_0
qhull=2020.2=h434a139_5
qt-main=5.15.2=hb6262e9_11
readline=8.2=h8228510_1
requests=2.32.3=pyhd8ed1ab_1
safetensors=0.4.5=py312hc50d6dc_1
scikit-image=0.25.0=py312h6a678d5_0
scikit-learn=1.5.2=py312h6a678d5_0
scipy=1.15.1=py312hc5e2394_0
segment-anything=1.0=pyhd8ed1ab_0
setuptools=75.8.0=pyhff2d567_0
sip=6.7.12=py312h6a678d5_1
six=1.17.0=pyhd8ed1ab_0
snappy=1.2.1=h8bd8927_1
sqlite=3.46.0=h6d4b2fc_0
sympy=1.13.3=pyh04b8f61_5
tbb=2021.13.0=hceb3a55_1
threadpoolctl=3.5.0=pyhc1e730c_0
tifffile=2025.2.18=py312h06a4308_0
tiledb=2.3.3=h77177df_3
timm=1.0.13=pyhd8ed1ab_0
tk=8.6.14=h39e8969_0
tomli=2.2.1=pyhd8ed1ab_1
torchaudio=2.5.1=py312_cu124
torchcam=0.4.0=pyhd8ed1ab_0
torchtriton=3.1.0=py312
torchvision=0.20.1=py312_cu124
tornado=6.4.2=py312h5eee18b_0
tqdm=4.67.1=pyhd8ed1ab_1
ttach=0.0.3=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_1
typing_extensions=4.12.2=pyha770c72_1
tzdata=2024b=hc8b5060_0
unicodedata2=15.1.0=py312h5eee18b_1
uriparser=0.9.8=hac33072_0
urllib3=2.3.0=pyhd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
x264=1!152.20180806=h14c3975_0
xerces-c=3.2.5=hac6953d_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libx11=1.8.10=h4f16b4b_1
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xz=5.6.3=hbcc6ac9_1
xz-gpl-tools=5.6.3=hbcc6ac9_1
xz-tools=5.6.3=hb9d3cd8_1
yaml=0.2.5=h7f98852_2
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h4ab18f5_6
zstandard=0.23.0=py312h2c38b39_1
zstd=1.5.6=ha6fb4c9_0
